\docType{package}
\name{bigramr-package}
\alias{bigramr}
\alias{bigramr-package}
\title{Tools to calculate unigram and bigram frequency counts of strings
  (e.g., words, pseudowords).}
\description{
  In psycho- and neurolinguistic work that uses written
  language stimuli, it is sometimes important to either
  control or manipulate certain properties of character
  strings, such as Bigram Frequency (cf. The English
  Lexicon Project online; Balota et al. 2007). This package
  offers tools to allow researchers to effortlessly
  calculate the Sum and Mean Unigram Frequencies and the
  Sum and Mean Bigram Frequencies of arbitrary character
  strings (such as words, pseudowords and non-words) based
  on five different case-sensitive unigram and bigram
  frequency English-language corpora. These corpora were
  collected and published by Jones & Mewhort (2004). The
  functions in this package will also work with
  user-supplied frequency corpora, thereby allowing its
  functionality to be extended to include other frequency
  corpora of English and other languages.
}
\author{
  Diogo Almeida \email{dalazal@gmail.com}
}
\references{
  Balota, D. A., Yap, M. J., Hutchison, K. A., Cortese, M.
  J., Kessler, B., Loftis, B., ... & Treiman, R. (2007).
  The English lexicon project. Behavior Research Methods,
  39(3), 445-459.

  Jones, M. N., & Mewhort, D. J. (2004). Case-sensitive
  letter and bigram frequency counts from large-scale
  English corpora. Behavior Research Methods, Instruments,
  & Computers, 36(3), 388-396.
}
\keyword{package}

